Zero-loss throughput and expdesign selection
--exp-design parameters enable to set the way to explore the
experimental design.

By default it is full-factorial (--exp-design full)

--exp-design zlt(RATE,PPS) enables a zero-loss-throughput search

Instead of going through all possible rates, it will look at the result
of PPS and do some heuristics + binary search to find the maximal
zero-loss throughput.
